Split math unit tests into multiple translation units
After adding tests for Complex<>, Visual Studio CI setup was consistently running out of heap memory.
Split the tests into separate translation units for float, double, complex float, complex double, and all ADL tests separately.
The split fixes this issue at the cost of slight code duplication.
